Start
The work is divided in two parts, represented by two folders.

Simple: It represents a simpler and more "engineering" approach to the problem.

Complex: The program in Complex try to recreate more precisely the process of the biocircuits implemented, modeling the different concentrations and gene expressions as variable and the promoters and repressors using guards that fire with a probability proportional to the promoter concentration.

The first circuit to be implement is "A synthetic multicellular system for
programmed pattern formation"
by Basu et al."